Add F-Droid build flavor.
We need a different build flavor for F-Droid because they forbid
auto-updaters as they download blobs from the internet (and it wouldn't
be possible to install the downloaded APK anyways because of the
signatures mismatch).

I chose to implement this using a build config field instead of using a
per-build-flavor source directory because I thought it introduces less
maintenance burden. In particular, per-build-flavor source directories
would require an interface in the main source directory for the update
manager, which would be implemented for each build flavor in a specific
source directory. Considering this project doesn't have CI builds yet, I
think it wouldn't be wise. Still we might reconsider should we need more
complex setup (such as different dependencies for the two build
flavors). The main downside of the build config field for now is that it
may be more prone to silent regressions.

Fixes #41